This is the final batch of commits from PostgreSQL 9.1 development, up
to the point where the REL9_1_STABLE branch was created, and 9.2
development started on the PostgreSQL master branch.

Notable upstream changes:

* Changes to "temporary" vs. "permanent" relation handling. Previously,
  relations were classified as either temporary or not-temporary. Now
  there is a concept of "relation persistence", which includes
  un-WAL-logged (but permanent) tables as a third option.

  Additionally, 9.1 adds the backend ID to the relpath of temporary
  tables. Because GPDB does not have a 1:1 relationship between backends
  and temp relations, this has been reverted for now and the concept of
  a special `TempRelBackendId` has been added. We may look into
  backporting pieces of the parallel session support from future
  versions of PostgreSQL to minimize this difference.

  There is currently a fatal flaw in our handling of relfilenodes for
  temporary relations: it is possible (though unlikely) for a temporary
  sequence and a permanent relation (or vice-versa) to point to the same
  shared buffer in the buffer cache. This may lead to corruption of that
  buffer. Our current attempt at a fix for this issue isn't good enough
  yet, so the offending code is marked with FIXMEs. We plan to solve this
  problem with a total overhaul of the way temporary relations are
  stored in the shared buffer cache.

* Collation is now supported per-column. Many plan nodes now have their
  own concept of input and output collation IDs. Notably, since this
  change involves heavy changes to generated plans, ORCA does not yet
  support non-default collation.

* The SERIALIZABLE isolation level was added. This is not currently
  supported in GPDB, because we don't have a way to analyze
  serialization conflicts in a distributed manner; we will "fall back"
  to REPEATABLE READ.

* GROUP BY now allows selection of columns in the grouping query without
  mentioning those columns in the GROUP BY clause, as long as those
  columns are from a table that's being grouped by its primary key. This
  requires some modification to the parallel grouping planner in GPDB.

* Support for INSTEAD OF triggers on views. Notably, this would allow
  "modification" of views, but GPDB currently has no support for this
  concept in the planner. This functionality is temporarily disabled for
  now.

* CTEs (WITH ... AS) can now contain data modification statements
  (INSERT/UPDATE/DELETE). This is not yet supported in GPDB.

* `make -k` is now supported for installcheck[-world] and several other
  Makefile targets.

* Foreign tables. Currently, GPDB does not support planning for
  statements made against foreign tables, so CREATE FOREIGN TABLE has
  been disabled.

* SELinux support and integration has been added with the SECURITY LABEL
  command and the associated sepgsql utility. SECURITY LABEL support is
  currently tested in GPDB via the upstream regression tests, but
  sepgsql is not.

* GUC API changes: assignment hooks are now split into two phases
  (check/assign).

Notable GPDB changes:

* Several pieces of the COPY dispatching code have been heavily
  refactored/rewritten to match upstream more closely. For now,
  compatibility with 5.x external table catalog entries has been
  maintained, though this is likely to change.

+++ contrib/pg_upgrade/relfilenode.c
@@ -3,8 +3,8 @@
  *
  *	relfilenode functions
  *
- *	Copyright (c) 2010, PostgreSQL Global Development Group
- *	$PostgreSQL: pgsql/contrib/pg_upgrade/relfilenode.c,v 1.8 2010/07/06 19:18:55 momjian Exp $
+ *	Copyright (c) 2010-2011, PostgreSQL Global Development Group
+ *	contrib/pg_upgrade/relfilenode.c
  */
 
 #include "pg_upgrade.h"
@@ -13,12 +13,11 @@
 #include "access/transam.h"
 
 
-static void transfer_single_new_db(migratorContext *ctx, pageCnvCtx *pageConverter,
+static void transfer_single_new_db(pageCnvCtx *pageConverter,
 					   FileNameMap *maps, int size);
-static void transfer_relfile(migratorContext *ctx, pageCnvCtx *pageConverter,
+static void transfer_relfile(pageCnvCtx *pageConverter,
 				 const char *fromfile, const char *tofile,
-				 const char *oldnspname, const char *oldrelname,
-				 const char *newnspname, const char *newrelname,
+				 const char *nspname, const char *relname,
 				 bool gpdb4_heap_conversion_needed,
 				 bool has_numerics, AttInfo *atts, int natts, RelType reltype);
 
@@ -32,45 +31,48 @@ char		scandir_file_pattern[MAXPGPATH];
  * physically link the databases.
  */
 const char *
-transfer_all_new_dbs(migratorContext *ctx, DbInfoArr *olddb_arr,
-					 DbInfoArr *newdb_arr, char *old_pgdata, char *new_pgdata)
+transfer_all_new_dbs(DbInfoArr *old_db_arr,
+				   DbInfoArr *new_db_arr, char *old_pgdata, char *new_pgdata)
 {
 	int			dbnum;
 	const char *msg = NULL;
 
-	prep_status(ctx, "Restoring user relation files");
+	prep_status("Restoring user relation files\n");
 
-	for (dbnum = 0; dbnum < newdb_arr->ndbs; dbnum++)
+	if (old_db_arr->ndbs != new_db_arr->ndbs)
+		pg_log(PG_FATAL, "old and new clusters have a different number of databases\n");
+
+	for (dbnum = 0; dbnum < old_db_arr->ndbs; dbnum++)
 	{
-		DbInfo	   *new_db = &newdb_arr->dbs[dbnum];
-		DbInfo	   *old_db = dbarr_lookup_db(olddb_arr, new_db->db_name);
+		DbInfo	   *old_db = &old_db_arr->dbs[dbnum];
+		DbInfo	   *new_db = &new_db_arr->dbs[dbnum];
 		FileNameMap *mappings;
 		int			n_maps;
 		pageCnvCtx *pageConverter = NULL;
 
-		if (!old_db)
-			pg_log(ctx, PG_FATAL,
-			   "the new cluster database %s was not found in the old cluster\n", new_db->db_name);
-		
+		if (strcmp(old_db->db_name, new_db->db_name) != 0)
+			pg_log(PG_FATAL, "old and new databases have different names: old \"%s\", new \"%s\"\n",
+				   old_db->db_name, new_db->db_name);
+
 		n_maps = 0;
-		mappings = gen_db_file_maps(ctx, old_db, new_db, &n_maps, old_pgdata,
+		mappings = gen_db_file_maps(old_db, new_db, &n_maps, old_pgdata,
 									new_pgdata);
 
 		if (n_maps)
 		{
-			print_maps(ctx, mappings, n_maps, new_db->db_name);
+			print_maps(mappings, n_maps, new_db->db_name);
 
 #ifdef PAGE_CONVERSION
-			msg = setupPageConverter(ctx, &pageConverter);
+			msg = setupPageConverter(&pageConverter);
 #endif
-			transfer_single_new_db(ctx, pageConverter, mappings, n_maps);
+			transfer_single_new_db(pageConverter, mappings, n_maps);
 
 			pg_free(mappings);
 		}
 	}
 
-	prep_status(ctx, "");		/* in case nothing printed */
-	check_ok(ctx);
+	prep_status("");			/* in case nothing printed */
+	check_ok();
 
 	return msg;
 }
@@ -83,13 +85,13 @@ transfer_all_new_dbs(migratorContext *ctx, DbInfoArr *olddb_arr,
  *	relfilenodes later in the upgrade process.
  */
 void
-get_pg_database_relfilenode(migratorContext *ctx, Cluster whichCluster)
+get_pg_database_relfilenode(ClusterInfo *cluster)
 {
-	PGconn	   *conn = connectToServer(ctx, "template1", whichCluster);
+	PGconn	   *conn = connectToServer(cluster, "template1");
 	PGresult   *res;
 	int			i_relfile;
 
-	res = executeQueryOrDie(ctx, conn,
+	res = executeQueryOrDie(conn,
 							"SELECT c.relname, c.relfilenode "
 							"FROM 	pg_catalog.pg_class c, "
 							"		pg_catalog.pg_namespace n "
@@ -99,10 +101,7 @@ get_pg_database_relfilenode(migratorContext *ctx, Cluster whichCluster)
 							"ORDER BY c.relname");
 
 	i_relfile = PQfnumber(res, "relfilenode");
-	if (whichCluster == CLUSTER_OLD)
-		ctx->old.pg_database_oid = atooid(PQgetvalue(res, 0, i_relfile));
-	else
-		ctx->new.pg_database_oid = atooid(PQgetvalue(res, 0, i_relfile));
+	cluster->pg_database_oid = atooid(PQgetvalue(res, 0, i_relfile));
 
 	PQclear(res);
 	PQfinish(conn);
@@ -115,25 +114,45 @@ get_pg_database_relfilenode(migratorContext *ctx, Cluster whichCluster)
  * create links for mappings stored in "maps" array.
  */
 static void
-transfer_single_new_db(migratorContext *ctx, pageCnvCtx *pageConverter,
+transfer_single_new_db(pageCnvCtx *pageConverter,
 					   FileNameMap *maps, int size)
 {
+	char		old_dir[MAXPGPATH];
+	struct dirent **namelist = NULL;
+	int			numFiles = 0;
 	int			mapnum;
+	int			fileno;
+
+	old_dir[0] = '\0';
 
 	for (mapnum = 0; mapnum < size; mapnum++)
 	{
 		char		old_file[MAXPGPATH];
 		char		new_file[MAXPGPATH];
-		struct dirent **namelist = NULL;
-		int			numFiles;
 		bool		seg0_missing;
 		struct stat st;
 
+		/* Changed tablespaces?  Need a new directory scan? */
+		if (strcmp(maps[mapnum].old_dir, old_dir) != 0)
+		{
+			if (numFiles > 0)
+			{
+				for (fileno = 0; fileno < numFiles; fileno++)
+					pg_free(namelist[fileno]);
+				pg_free(namelist);
+			}
+
+			snprintf(old_dir, sizeof(old_dir), "%s", maps[mapnum].old_dir);
+			numFiles = pg_scandir(old_dir, &namelist, NULL);
+		}
+
 		/* Copying files might take some time, so give feedback. */
 
-		snprintf(old_file, sizeof(old_file), "%s/%u", maps[mapnum].old_file, maps[mapnum].old);
-		snprintf(new_file, sizeof(new_file), "%s/%u", maps[mapnum].new_file, maps[mapnum].new);
-		pg_log(ctx, PG_REPORT, OVERWRITE_MESSAGE, old_file);
+		snprintf(old_file, sizeof(old_file), "%s/%u", maps[mapnum].old_dir,
+				 maps[mapnum].old_relfilenode);
+		snprintf(new_file, sizeof(new_file), "%s/%u", maps[mapnum].new_dir,
+				 maps[mapnum].new_relfilenode);
+		pg_log(PG_REPORT, OVERWRITE_MESSAGE, old_file);
 
 		/*
 		 * Copy/link the relation file to the new cluster
@@ -142,9 +161,8 @@ transfer_single_new_db(migratorContext *ctx, pageCnvCtx *pageConverter,
 
 		if (!maps[mapnum].missing_seg0_ok || stat(old_file, &st) == 0)
 		{
-			transfer_relfile(ctx, pageConverter, old_file, new_file,
-						 maps[mapnum].old_nspname, maps[mapnum].old_relname,
-						 maps[mapnum].new_nspname, maps[mapnum].new_relname,
+			transfer_relfile(pageConverter, old_file, new_file,
+						 maps[mapnum].nspname, maps[mapnum].relname,
 						 maps[mapnum].gpdb4_heap_conversion_needed,
 						 maps[mapnum].has_numerics, maps[mapnum].atts,
 						 maps[mapnum].natts, maps[mapnum].type);
@@ -154,31 +172,30 @@ transfer_single_new_db(migratorContext *ctx, pageCnvCtx *pageConverter,
 			seg0_missing = true;
 
 		/* fsm/vm files added in PG 8.4 */
-		if (GET_MAJOR_VERSION(ctx->old.major_version) >= 804)
+		if (GET_MAJOR_VERSION(old_cluster.major_version) >= 804)
 		{
 			/*
-			 * Now copy/link any fsm and vm files, if they exist
+			 * Copy/link any fsm and vm files, if they exist
 			 */
-			snprintf(scandir_file_pattern, sizeof(scandir_file_pattern), "%u_", maps[mapnum].old);
-			numFiles = pg_scandir(ctx, maps[mapnum].old_file, &namelist, dir_matching_filenames);
+			snprintf(scandir_file_pattern, sizeof(scandir_file_pattern), "%u_",
+					 maps[mapnum].old_relfilenode);
 
-			while (numFiles--)
+			for (fileno = 0; fileno < numFiles; fileno++)
 			{
-				snprintf(old_file, sizeof(old_file), "%s/%s", maps[mapnum].old_file,
-						 namelist[numFiles]->d_name);
-				snprintf(new_file, sizeof(new_file), "%s/%u%s", maps[mapnum].new_file,
-				  maps[mapnum].new, strchr(namelist[numFiles]->d_name, '_'));
+				if (strncmp(namelist[fileno]->d_name, scandir_file_pattern,
+							strlen(scandir_file_pattern)) == 0)
+				{
+					snprintf(old_file, sizeof(old_file), "%s/%s", maps[mapnum].old_dir,
+							 namelist[fileno]->d_name);
+					snprintf(new_file, sizeof(new_file), "%s/%u%s", maps[mapnum].new_dir,
+							 maps[mapnum].new_relfilenode, strchr(namelist[fileno]->d_name, '_'));
 
-				unlink(new_file);
-				transfer_relfile(ctx, pageConverter, old_file, new_file,
-						  maps[mapnum].old_nspname, maps[mapnum].old_relname,
-								 maps[mapnum].new_nspname, maps[mapnum].new_relname,
-								 false, false, NULL, 0, FSM);
-
-				pg_free(namelist[numFiles]);
+					unlink(new_file);
+					transfer_relfile(pageConverter, old_file, new_file,
+									 maps[mapnum].nspname, maps[mapnum].relname,
+									 false, false, NULL, 0, FSM);
+				}
 			}
-
-			pg_free(namelist);
 		}
 
 		/*
@@ -188,27 +205,34 @@ transfer_single_new_db(migratorContext *ctx, pageCnvCtx *pageConverter,
 		 * relfilenode.3, ...  'fsm' and 'vm' files use underscores so are not
 		 * copied.
 		 */
-		snprintf(scandir_file_pattern, sizeof(scandir_file_pattern), "%u.", maps[mapnum].old);
-		numFiles = pg_scandir(ctx, maps[mapnum].old_file, &namelist, dir_matching_filenames);
+		snprintf(scandir_file_pattern, sizeof(scandir_file_pattern), "%u.",
+				 maps[mapnum].old_relfilenode);
 
-		while (numFiles--)
+		for (fileno = 0; fileno < numFiles; fileno++)
 		{
-			snprintf(old_file, sizeof(old_file), "%s/%s", maps[mapnum].old_file,
-					 namelist[numFiles]->d_name);
-			snprintf(new_file, sizeof(new_file), "%s/%u%s", maps[mapnum].new_file,
-				  maps[mapnum].new, strchr(namelist[numFiles]->d_name, '.'));
+			if (strncmp(namelist[fileno]->d_name, scandir_file_pattern,
+						strlen(scandir_file_pattern)) == 0)
+			{
+				snprintf(old_file, sizeof(old_file), "%s/%s", maps[mapnum].old_dir,
+						 namelist[fileno]->d_name);
+				snprintf(new_file, sizeof(new_file), "%s/%u%s", maps[mapnum].new_dir,
+						 maps[mapnum].new_relfilenode, strchr(namelist[fileno]->d_name, '.'));
 
-			unlink(new_file);
-			transfer_relfile(ctx, pageConverter, old_file, new_file,
-						  maps[mapnum].old_nspname, maps[mapnum].old_relname,
-							 maps[mapnum].new_nspname, maps[mapnum].new_relname,
-							 maps[mapnum].gpdb4_heap_conversion_needed,
-							 maps[mapnum].has_numerics, maps[mapnum].atts,
-							 maps[mapnum].natts, maps[mapnum].type);
-
-			pg_free(namelist[numFiles]);
+				unlink(new_file);
+				transfer_relfile(pageConverter, old_file, new_file,
+								 maps[mapnum].nspname, maps[mapnum].relname,
+								 maps[mapnum].gpdb4_heap_conversion_needed,
+								 maps[mapnum].has_numerics, maps[mapnum].atts,
+								 maps[mapnum].natts, maps[mapnum].type);
+			}
 		}
+	}
 
+
+	if (numFiles > 0)
+	{
+		for (fileno = 0; fileno < numFiles; fileno++)
+			pg_free(namelist[fileno]);
 		pg_free(namelist);
 	}
 }
@@ -220,9 +244,8 @@ transfer_single_new_db(migratorContext *ctx, pageCnvCtx *pageConverter,
  * Copy or link file from old cluster to new one.
  */
 static void
-transfer_relfile(migratorContext *ctx, pageCnvCtx *pageConverter, const char *oldfile,
-		 const char *newfile, const char *oldnspname, const char *oldrelname,
-				 const char *newnspname, const char *newrelname,
+transfer_relfile(pageCnvCtx *pageConverter, const char *old_file,
+				 const char *new_file, const char *nspname, const char *relname,
 				 bool gpdb4_heap_conversion_needed,
 				 bool has_numerics, AttInfo *atts, int natts, RelType type)
 {
@@ -230,46 +253,44 @@ transfer_relfile(migratorContext *ctx, pageCnvCtx *pageConverter, const char *ol
 
 	if (gpdb4_heap_conversion_needed)
 	{
-		pg_log(ctx, PG_INFO, "copying and converting %s to %s\n", oldfile, newfile);
+		pg_log(PG_INFO, "copying and converting %s to %s\n", old_file, new_file);
 
-		if ((msg = convert_gpdb4_heap_file(ctx, oldfile, newfile,
+		if ((msg = convert_gpdb4_heap_file(old_file, new_file,
 										   has_numerics, atts, natts)) != NULL)
-			pg_log(ctx, PG_FATAL, "error while copying %s.%s(%s) to %s.%s(%s): %s\n",
-				   oldnspname, oldrelname, oldfile, newnspname, newrelname, newfile, msg);
+			pg_log(PG_FATAL, "error while copying %s.%s (%s to %s): %s\n",
+				   nspname, relname, old_file, new_file, msg);
 
 		return;
 	}
 
-	if ((ctx->transfer_mode == TRANSFER_MODE_LINK) && (pageConverter != NULL))
-		pg_log(ctx, PG_FATAL, "this migration requires page-by-page conversion, "
+	if ((user_opts.transfer_mode == TRANSFER_MODE_LINK) && (pageConverter != NULL))
+		pg_log(PG_FATAL, "this upgrade requires page-by-page conversion, "
 			   "you must use copy-mode instead of link-mode\n");
 
-	if (ctx->transfer_mode == TRANSFER_MODE_COPY)
+	if (user_opts.transfer_mode == TRANSFER_MODE_COPY)
 	{
-		if (ctx->checksum_mode != CHECKSUM_NONE && type == HEAP)
+		if (user_opts.checksum_mode != CHECKSUM_NONE && type == HEAP)
 		{
-			pg_log(ctx, PG_INFO, "rewriting \"%s\" to \"%s\"\n",
-				   oldfile, newfile);
-			rewriteHeapPageChecksum(ctx, oldfile, newfile, newnspname, newrelname);
+			pg_log(PG_INFO, "rewriting \"%s\" to \"%s\"\n", old_file, new_file);
+			rewriteHeapPageChecksum(old_file, new_file, nspname, relname);
 		}
 		else
 		{
-			pg_log(ctx, PG_INFO, "copying \"%s\" to \"%s\"\n",
-				   oldfile, newfile);
-			if ((msg = copyAndUpdateFile(ctx, pageConverter, oldfile, newfile, true)) != NULL)
-				pg_log(ctx, PG_FATAL, "error while copying %s.%s(%s) to %s.%s(%s): %s\n",
-					   oldnspname, oldrelname, oldfile, newnspname, newrelname, newfile, msg);
+			pg_log(PG_INFO, "copying %s to %s\n", old_file, new_file);
+
+			if ((msg = copyAndUpdateFile(pageConverter, old_file, new_file, true)) != NULL)
+				pg_log(PG_FATAL, "error while copying %s.%s (%s to %s): %s\n",
+					   nspname, relname, old_file, new_file, msg);
 		}
 	}
 	else
 	{
-		pg_log(ctx, PG_INFO, "linking %s to %s\n", oldfile, newfile);
+		pg_log(PG_INFO, "linking %s to %s\n", old_file, new_file);
 
-		if ((msg = linkAndUpdateFile(ctx, pageConverter, oldfile, newfile)) != NULL)
-			pg_log(ctx, PG_FATAL,
-			   "error while creating link from %s.%s(%s) to %s.%s(%s): %s\n",
-				   oldnspname, oldrelname, oldfile, newnspname, newrelname,
-				   newfile, msg);
+		if ((msg = linkAndUpdateFile(pageConverter, old_file, new_file)) != NULL)
+			pg_log(PG_FATAL,
+				   "error while creating link from %s.%s (%s to %s): %s\n",
+				   nspname, relname, old_file, new_file, msg);
 	}
 	return;
 }
